+++ libevmasm/Assembly.h
@@ -151,9 +151,20 @@ public:
 
 	/// Create a JSON representation of the assembly.
 	Json::Value assemblyJSON(
-		std::map<std::string, unsigned> const& _sourceIndices = std::map<std::string, unsigned>(),
+		std::map<std::string, unsigned> const& _sourceIndices = {},
 		bool _includeSourceList = true
 	) const;
+
+	/// Loads the JSON representation of assembly.
+	/// @param _json JSON object containing assembly in the format produced by assemblyJSON().
+	/// @param _sourceList list of source names.
+	/// @param _level this function might be called recursively, _level reflects the nesting level.
+	/// @returns Resulting Assembly object loaded from given json.
+	static std::shared_ptr<Assembly> fromJSON(
+		Json::Value const& _json,
+		std::vector<std::string> const& _sourceList = {},
+		int _level = 0
+	);
 
 	/// Mark this assembly as invalid. Calling ``assemble`` on it will throw.
 	void markAsInvalid() { m_invalid = true; }
@@ -170,6 +181,18 @@ protected:
 	std::map<u256, u256> const& optimiseInternal(OptimiserSettings const& _settings, std::set<size_t> _tagsReferencedFromOutside);
 
 	unsigned codeSize(unsigned subTagSize) const;
+
+	/// Add all assembly items from given JSON array. This function imports the items by iterating through
+	/// the code array. This method only works on clean Assembly objects that don't have any items defined yet.
+	/// @param _json JSON array that contains assembly items (e.g. json['.code'])
+	/// @param _sourceList list of source names.
+	void importAssemblyItemsFromJSON(Json::Value const& _code, std::vector<std::string> const& _sourceList);
+
+	/// Creates an AssemblyItem from a given JSON representation.
+	/// @param _json JSON object that consists a single assembly item
+	/// @param _sourceList list of source names.
+	/// @returns AssemblyItem of _json argument.
+	AssemblyItem createAssemblyItemFromJSON(Json::Value const& _json, std::vector<std::string> const& _sourceList);
 
 private:
 	bool m_invalid = false;